Split state to game and manager. Add more implementation for the link sharing feature
