+++ src/components/message.rs
@@ -1,6 +1,6 @@
 use yew::prelude::*;
 
-use crate::state::GameMode;
+use crate::manager::GameMode;
 use crate::Msg as GameMsg;
 
 const FORMS_LINK_TEMPLATE_ADD: &str = "https://docs.google.com/forms/d/e/1FAIpQLSfH8gs4sq-Ynn8iGOvlc99J_zOG2rJEC4m8V0kCgF_en3RHFQ/viewform?usp=pp_url&entry.461337706=Lis%C3%A4yst%C3%A4&entry.560255602=";
@@ -13,6 +13,7 @@ pub struct Props {
     pub is_unknown: bool,
     pub is_winner: bool,
     pub is_guessing: bool,
+    pub is_hidden: bool,
     pub word: String,
     pub last_guess: String,
     pub game_mode: GameMode,
@@ -63,7 +64,31 @@ impl Component for Message {
             <div class="message">
                 { &props.message }
                 <div class="message-small">{{
-                    if props.is_unknown {
+                    if props.is_hidden {
+                        let callback = props.callback.clone();
+                        let reveal_hidden_tiles = ctx.link().callback(move |e: MouseEvent| {
+                            e.prevent_default();
+                            callback.emit(GameMsg::RevealHiddenTiles);
+                            Msg::SetIsEmojisCopied(true)
+                        });
+                        let callback = props.callback.clone();
+                        let reset_game = Callback::from(move |e: MouseEvent| {
+                            e.prevent_default();
+                            callback.emit(GameMsg::ResetGame);
+                        });
+
+                        html! {
+                            <>
+                                <a class="link" href={"javascript:void(0)"} onclick={reveal_hidden_tiles}>
+                                    {"Paljasta ratkaisu"}
+                                </a>
+                                {" | "}
+                                <a class="link" href={"javascript:void(0)"} onclick={reset_game}>
+                                    {"Kokeile itse"}
+                                </a>
+                            </>
+                        }
+                    } else if props.is_unknown {
                         let last_guess = props.last_guess.to_lowercase();
                         html! {
                             <a class="link" href={format!("{}{}", FORMS_LINK_TEMPLATE_ADD, last_guess)}
