+++ src/main.rs
@@ -6,7 +6,8 @@ use yew::prelude::*;
 
 mod components;
 mod migration;
-mod state;
+mod manager;
+mod game;
 
 use components::{
     board::Board,
@@ -14,7 +15,8 @@ use components::{
     keyboard::Keyboard,
     modal::{HelpModal, MenuModal},
 };
-use state::{Game, GameMode, State, Theme, TileState, WordList};
+use manager::{GameMode, Manager, Theme, TileState, WordList};
+use game::{Game};
 
 const ALLOWED_KEYS: [char; 28] = [
     'Q', 'W', 'E', 'R', 'T', 'Y', 'U', 'I', 'O', 'P', 'A', 'S', 'D', 'F', 'G', 'H', 'J', 'K', 'L',
@@ -37,10 +39,12 @@ pub enum Msg {
     ChangeTheme(Theme),
     ShareEmojis,
     ShareLink,
+    RevealHiddenTiles,
+    ResetGame,
 }
 
 pub struct App {
-    state: State,
+    manager: Manager,
     is_help_visible: bool,
     is_menu_visible: bool,
     keyboard_listener: Option<Closure<dyn Fn(KeyboardEvent)>>,
@@ -52,7 +56,7 @@ impl Component for App {
 
     fn create(_ctx: &Context<Self>) -> Self {
         Self {
-            state: State::new(),
+            manager: Manager::new(),
             is_help_visible: false,
             is_menu_visible: false,
             keyboard_listener: None,
@@ -107,13 +111,13 @@ impl Component for App {
 
     fn update(&mut self, ctx: &Context<Self>, msg: Self::Message) -> bool {
         match msg {
-            Msg::KeyPress(c) => self.state.game.push_character(c),
-            Msg::Backspace => self.state.game.pop_character(),
+            Msg::KeyPress(c) => self.manager.game.push_character(c),
+            Msg::Backspace => self.manager.game.pop_character(),
             Msg::Enter => {
                 let link = ctx.link();
 
-                if !self.state.game.is_guessing {
-                    if let GameMode::DailyWord(_) = self.state.game.game_mode {
+                if !self.manager.game.is_guessing {
+                    if matches!(self.manager.game.game_mode, GameMode::DailyWord(_) | GameMode::Shared) {
                         link.send_message(Msg::ChangePreviousGameMode);
                     } else {
                         link.send_message(Msg::NextWord);
@@ -124,8 +128,8 @@ impl Component for App {
 
                 true
             }
-            Msg::Guess => self.state.submit_guess(),
-            Msg::NextWord => self.state.game.next_word(),
+            Msg::Guess => self.manager.submit_guess(),
+            Msg::NextWord => self.manager.game.next_word(),
             Msg::ToggleHelp => {
                 self.is_help_visible = !self.is_help_visible;
                 self.is_menu_visible = false;
@@ -137,43 +141,42 @@ impl Component for App {
                 true
             }
             Msg::ChangeWordLength(new_length) => {
-                self.state.change_word_length(new_length);
+                self.manager.change_word_length(new_length);
                 self.is_menu_visible = false;
                 self.is_help_visible = false;
                 true
             }
             Msg::ChangeGameMode(new_mode) => {
-                self.state.change_game_mode(new_mode);
+                self.manager.change_game_mode(new_mode);
                 self.is_menu_visible = false;
                 self.is_help_visible = false;
                 true
             }
             Msg::ChangeWordList(new_list) => {
-                self.state.change_word_list(new_list);
+                self.manager.change_word_list(new_list);
                 self.is_menu_visible = false;
                 self.is_help_visible = false;
                 true
             }
             Msg::ChangePreviousGameMode => {
-                self.state.change_previous_game_mode();
+                self.manager.change_previous_game_mode();
                 true
             }
             Msg::ChangeAllowProfanities(is_allowed) => {
-                self.state.change_allow_profanities(is_allowed);
+                self.manager.change_allow_profanities(is_allowed);
                 self.is_menu_visible = false;
                 self.is_help_visible = false;
                 true
             }
             Msg::ChangeTheme(theme) => {
-                self.state.change_theme(theme);
-                true
+                self.manager.change_theme(theme)
             }
             Msg::ShareEmojis => {
                 #[cfg(web_sys_unstable_apis)]
                 {
                     use web_sys::Navigator;
 
-                    let emojis = self.state.share_emojis();
+                    let emojis = self.manager.share_emojis();
                     let window: Window = window().expect("window not available");
                     let navigator: Navigator = window.navigator();
                     if let Some(clipboard) = navigator.clipboard() {
@@ -187,7 +190,7 @@ impl Component for App {
                 {
                     use web_sys::Navigator;
 
-                    if let Some(link) = self.state.share_link() {
+                    if let Some(link) = self.manager.share_link() {
                         let window: Window = window().expect("window not available");
                         let navigator: Navigator = window.navigator();
                         if let Some(clipboard) = navigator.clipboard() {
@@ -197,7 +200,13 @@ impl Component for App {
                     }
                 }
                 true
-            }
+            },
+            Msg::RevealHiddenTiles => {
+                self.manager.reveal_hidden_tiles()
+            },
+            Msg::ResetGame => {
+                self.manager.reset_game()
+            },
         }
     }
 
@@ -206,12 +215,12 @@ impl Component for App {
 
         let keyboard_state = ALLOWED_KEYS
             .iter()
-            .map(|key| (*key, self.state.game.keyboard_tilestate(key)))
+            .map(|key| (*key, self.manager.game.keyboard_tilestate(key)))
             .collect::<HashMap<char, TileState>>();
 
-        let word = self.state.game.word.iter().collect::<String>();
+        let word = self.manager.game.word.iter().collect::<String>();
 
-        let last_guess = self.state.game.guesses[self.state.game.current_guess]
+        let last_guess = self.manager.game.guesses[self.manager.game.current_guess]
             .iter()
             .map(|(c, _)| c)
             .collect::<String>();
@@ -219,34 +228,34 @@ impl Component for App {
         let today = Local::now().naive_local().date();
 
         html! {
-            <div class={classes!("game", self.state.theme.to_string())}>
+            <div class={classes!("game", self.manager.theme.to_string())}>
                 <Header
                     on_toggle_help_cb={link.callback(|_| Msg::ToggleHelp)}
                     on_toggle_menu_cb={link.callback(|_| Msg::ToggleMenu)}
-                    streak={self.state.game.streak}
-                    game_mode={self.state.game.game_mode}
+                    streak={self.manager.game.streak}
+                    game_mode={self.manager.game.game_mode}
                     daily_word_number={Game::get_daily_word_index(today) + 1}
                 />
 
                 <Board
-                    is_guessing={self.state.game.is_guessing}
-                    is_reset={self.state.game.is_reset}
-                    is_hidden={self.state.game.is_hidden}
-                    guesses={self.state.game.guesses.clone()}
-                    previous_guesses={self.state.game.previous_guesses.clone()}
-                    current_guess={self.state.game.current_guess}
-                    max_guesses={self.state.game.max_guesses}
-                    word_length={self.state.game.word_length}
+                    is_guessing={self.manager.game.is_guessing}
+                    is_reset={self.manager.game.is_reset}
+                    is_hidden={self.manager.game.is_hidden}
+                    guesses={self.manager.game.guesses.clone()}
+                    previous_guesses={self.manager.game.previous_guesses.clone()}
+                    current_guess={self.manager.game.current_guess}
+                    max_guesses={self.manager.game.max_guesses}
+                    word_length={self.manager.game.word_length}
                 />
 
                 <Keyboard
                     callback={link.callback(move |msg| msg)}
-                    is_unknown={self.state.game.is_unknown}
-                    is_winner={self.state.game.is_winner}
-                    is_guessing={self.state.game.is_guessing}
-                    is_hidden={self.state.game.is_hidden}
-                    game_mode={self.state.game.game_mode}
-                    message={self.state.game.message.clone()}
+                    is_unknown={self.manager.game.is_unknown}
+                    is_winner={self.manager.game.is_winner}
+                    is_guessing={self.manager.game.is_guessing}
+                    is_hidden={self.manager.game.is_hidden}
+                    game_mode={self.manager.game.game_mode}
+                    message={self.manager.game.message.clone()}
                     word={word}
                     last_guess={last_guess}
                     keyboard={keyboard_state}
@@ -265,14 +274,14 @@ impl Component for App {
                         html! {
                             <MenuModal
                                 callback={link.callback(move |msg| msg)}
-                                game_mode={self.state.current_game_mode}
-                                word_length={self.state.current_word_length}
-                                current_word_list={self.state.current_word_list}
-                                allow_profanities={self.state.allow_profanities}
-                                theme={self.state.theme}
-                                max_streak={self.state.max_streak}
-                                total_played={self.state.total_played}
-                                total_solved={self.state.total_solved}
+                                game_mode={self.manager.current_game_mode}
+                                word_length={self.manager.current_word_length}
+                                current_word_list={self.manager.current_word_list}
+                                allow_profanities={self.manager.allow_profanities}
+                                theme={self.manager.theme}
+                                max_streak={self.manager.max_streak}
+                                total_played={self.manager.total_played}
+                                total_solved={self.manager.total_solved}
                             />
                         }
                     } else {
